pyproject.toml
[project]
name = "psql-update"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ichikawa" }
]
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["uv_build>=0.9.16,<0.10.0"]
build-backend = "uv_build"
